internal: Move `translate_regex` into dialect
We actually can't do what I'd hoped re other mysql & sqlite, but this prepares the ground for when that's available

It's also the first `translate_*` to move into the DialectHandler. I think one approach is that DialectHandler does _all_ translations. But no need to start on that until it's required
